+++ otelconf/x/config_test.go
@@ -2284,11 +2284,14 @@ func TestUnmarshalResourceJson(t *testing.T) {
 	}{
 		{
 			name:       "valid with all detectors",
-			jsonConfig: []byte(`{"detection/development": {"detectors": [{"aws.eks": null},{"container": null},{"host": null},{"process": null},{"service": null}]}}`),
-			yamlConfig: []byte("detection/development:\n  detectors:\n    - aws.eks:\n    - container:\n    - host:\n    - process:\n    - service:"),
+			jsonConfig: []byte(`{"detection/development": {"detectors": [{"gcp": null},{"aws.eks": null},{"container": null},{"host": null},{"process": null},{"service": null}]}}`),
+			yamlConfig: []byte("detection/development:\n  detectors:\n    - gcp:\n    - aws.eks:\n    - container:\n    - host:\n    - process:\n    - service:"),
 			wantResource: Resource{
 				DetectionDevelopment: &ExperimentalResourceDetection{
 					Detectors: []ExperimentalResourceDetector{
+						{
+							GCP: ExperimentalGCPResourceDetector{},
+						},
 						{
 							AWSEKS: ExperimentalAWSEKSResourceDetector{},
 						},
@@ -2310,11 +2313,14 @@ func TestUnmarshalResourceJson(t *testing.T) {
 		},
 		{
 			name:       "valid non-nil with all detectors",
-			jsonConfig: []byte(`{"detection/development": {"detectors": [{"aws.eks": {}},{"container": {}},{"host": {}},{"process": {}},{"service": {}}]}}`),
-			yamlConfig: []byte("detection/development:\n  detectors:\n    - aws.eks: {}\n    - container: {}\n    - host: {}\n    - process: {}\n    - service: {}"),
+			jsonConfig: []byte(`{"detection/development": {"detectors": [{"gcp": {}},{"aws.eks": {}},{"container": {}},{"host": {}},{"process": {}},{"service": {}}]}}`),
+			yamlConfig: []byte("detection/development:\n  detectors:\n    - gcp: {}\n    - aws.eks: {}\n    - container: {}\n    - host: {}\n    - process: {}\n    - service: {}"),
 			wantResource: Resource{
 				DetectionDevelopment: &ExperimentalResourceDetection{
 					Detectors: []ExperimentalResourceDetector{
+						{
+							GCP: ExperimentalGCPResourceDetector{},
+						},
 						{
 							AWSEKS: ExperimentalAWSEKSResourceDetector{},
 						},
@@ -2333,6 +2339,19 @@ func TestUnmarshalResourceJson(t *testing.T) {
 					},
 				},
 			},
+		},
+		{
+			name:       "invalid gcp detector",
+			jsonConfig: []byte(`{"detection/development": {"detectors": [{"gcp": 1}]}}`),
+			yamlConfig: []byte("detection/development:\n  detectors:\n    - gcp: 1"),
+			wantResource: Resource{
+				DetectionDevelopment: &ExperimentalResourceDetection{
+					Detectors: []ExperimentalResourceDetector{
+						{},
+					},
+				},
+			},
+			wantErrT: newErrUnmarshal(&ExperimentalResourceDetector{}),
 		},
 		{
 			name:       "invalid aws eks detector",